AAI-59 Add info button next to add annotations button
